Refactor: [BADA-74] inputField 폴더 명 수정 #21
